Fix the compile errors the cancelled CI runs never caught
Every previous push on this branch had its CI run cancelled as
superseded the moment the next commit landed, so several commits were
never compiled by anyone. The first run that was allowed to finish
found all of them at once; this is that list.

- SettingsScreen's caller passed linuxUserspace by name after a
  positional run, but the parameter is second - every positional
  argument after it slid one slot. Passed positionally now.
- LinuxUserspaceSection had @Composable twice (one before its KDoc, one
  after), and SettingsSection had lost its @Composable entirely - the
  composable helpers at the end of the file only made sense as cascade.
- The section's `when (state)` judged a nullable state without a null
  branch; supported implies a state, and the branch says that.
- android.os.Process has no myGid() in the modern SDK - the gid comes
  from Os.getgid(), falling back to the uid, which on Android is the
  same number for an app.
- ByteArray has no flatten(); the captured command output is assembled
  with one exact-size copy.
- The tar linkFlag comparisons needed bytes, and workspace fileCount
  counts an Int into a Long.
- RootfsInstaller's progress callbacks are not suspending: the
  downloader's chunk callback cannot call a suspend onProgress, so the
  progress type is a plain function and cancellation is checked through
  the captured Job instead.
- openedAtMs belongs on the TerminalChannel interface (the liveness
  probe reads it through the interface type), so both channels
  implement it - the local channel records when its pty was forked.

+++ app/src/main/java/dev/eclipse/ssh/linux/LinuxWorkspaceManager.kt
@@ -33,7 +33,7 @@ class LinuxWorkspaceManager(
     /** Total size on disk, for the settings screen's storage line. Walks the tree; call off the UI thread. */
     fun sizeBytes(): Long = files().sumOf { it.length() }
 
-    fun fileCount(): Long = files().count()
+    fun fileCount(): Long = files().count().toLong()
 
     private fun files(): Sequence<File> =
         if (workspaceDir.isDirectory) workspaceDir.walkTopDown().filter { it.isFile } else emptySequence()
@@ -97,7 +97,7 @@ class LinuxWorkspaceManager(
                             java.nio.file.Path.of(entry.linkName),
                         )
                     }
-                    TarArchiveEntry.LF_NORMAL, 0 -> {
+                    TarArchiveEntry.LF_NORMAL, 0.toByte() -> {
                         target.parentFile?.mkdirs()
                         target.outputStream().use { output -> tar.copyTo(output) }
                         target.setExecutable((entry.mode and 0b001_001_001) != 0, true)
